Add writable checks.

+++ share/julia-install/julia/os/ubuntu/setup.sh
@@ -9,10 +9,13 @@ systemctl mask apt-daily.service;
 systemctl mask apt-daily-upgrade.service;
 systemctl daemon-reload;
 
-# Archive Apt settings
-cp /etc/apt/apt.conf.d/10periodic /etc/apt/apt.conf.d/10periodic.ji-bak
-# Disable periodic activities of apt to be safe
-cat <<EOF >/etc/apt/apt.conf.d/10periodic;
+# Archive Apt settings and disable periodic updates if we can
+if [ -w /etc/apt/apt.conf.d/10periodic ]
+then
+  # File has write permission so we can restore a backup.
+  cp /etc/apt/apt.conf.d/10periodic /tmp/10periodic.ji-bak
+  # Disable periodic activities of apt to be safe
+  cat <<EOF >/etc/apt/apt.conf.d/10periodic;
 APT::Periodic::Enable "0";
 APT::Periodic::Update-Package-Lists "0";
 APT::Periodic::Download-Upgradeable-Packages "0";
@@ -20,8 +23,15 @@ APT::Periodic::AutocleanInterval "0";
 APT::Periodic::Unattended-Upgrade "0";
 EOF
 
+else
+  echo "You don't have write permission for /etc/apt/apt.conf.d/10periodic"
+fi
+
 # Cleanup
-rm -rf /var/log/unattended-upgrades;
+if [ -w /etc/apt/apt.conf.d/10periodic ]
+then
+  rm -rf /var/log/unattended-upgrades;
+fi
 apt-get --assume-yes purge unattended-upgrades;
 
 echo `ps faux | grep [a]pt`
@@ -32,5 +42,11 @@ pkgs='build-essential cmake gfortran libatomic1 m4 perl pkg-config python wget'
 ${sudo} apt-get install ${opts} ${pkgs}
 
 # Restore Apt settings
-rm -rf /etc/apt/apt.conf.d/10periodic
-mv /etc/apt/apt.conf.d/10periodic.ji-bak /etc/apt/apt.conf.d/10periodic
+if [ -w /etc/apt/apt.conf.d/10periodic ]
+then
+  # File has write permission
+  rm -rf /etc/apt/apt.conf.d/10periodic
+  mv /tmp/10periodic.ji-bak /etc/apt/apt.conf.d/10periodic
+else
+  echo "You don't have write permission for /etc/apt/apt.conf.d/10periodic"
+fi
